fix(demo): provide cached svg for codepen
Codepen will attempt to make an http request for assets causing a CORS
issue.  This is a problem since the assets are served from the docs site
instead of a CDN.  If the assets were served from the CDN, we could
configure the server to Access-Control-Allow-Origin' => '*', which would
allow codepen to access the assets.  This is a temporary solution until
the CDN is put in place.